Enemy AOE casting circle
Made enemy AOE casting circle with damage and a progress bar. It works well. Also has changes to the enemy boss sprite.

A huge step has been made in terms of the AOE. It really helped to stop fretting about the details such as the map or the magic circle and the looks. wah